create new folder within components with row component
